+++ src/app/basement/restricted_page/restricted_page.js
@@ -1,0 +1,7 @@
+"use strict";
+import restrictedPageTemplate from "./restricted_page.html";
+import app from "../../app";
+
+app.component("restrictedPage", {
+  template: restrictedPageTemplate,
+});