+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Resolved #919: Add color picker for navigation color field.
 - Fixed #934: Install `rsync` and check for its binary in CI.
 - Fixed #907: Handle user role permission.
+- Fixed #937: Bigger editor font size.
 
 
 2018.7.9