+++ CHANGELOG.md
@@ -5,7 +5,9 @@
 - Fixed #877: Fix uploading images in editor via Firefox.
 - Fixed #399: Filter image selection for images only via backend.
 - Resolved #919: Add color picker for navigation color field.
-- Fixed #934: Install `rsync` and check for its binary in CI. 
+- Fixed #934: Install `rsync` and check for its binary in CI.
+- Fixed #907: Handle user role permission.
+- Fixed #937: Bigger editor font size.
 
 
 2018.7.9