+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fixed #907: Handle user role permission.
 - Fixed #937: Bigger editor font size.
 - Fixed #915: Add live post preview from dashboard editor.
+- Fixed #953: Fix when pasting content to editor the page scrolls to the top.
 
 
 2018.7.9